♻️Refactor: QR SCAN 페이지 리팩토링
QR SCAN 페이지 리팩토링
