+++ src/Page/signupProcessPage/ManagerFirstPage.jsx
@@ -7,13 +7,9 @@ import {
   Box,
   Button,
   Checkbox,
-  FormControl,
   FormControlLabel,
-  FormLabel,
   Input,
   MenuItem,
-  Radio,
-  RadioGroup,
   Select,
   TextField,
   Typography,
