+++ src/constant/qr.ts
@@ -1,1 +1,5 @@
 export const QRSCAN_PERIOD = 3000;
+
+export const QRSCAN_NOT_VALID = "QR 코드가 유효하지 않습니다.";
+
+export const QRSCAN_FAILED = "QR 스캐너를 시작하는데 실패했습니다.";
